Replace clumsy Sentence abstraction with ItemSentenceMapping. move mapping algos to agg_surprisals. adjust tests accordingly.
